Decompile func0f01bea0

+++ src/include/game/data/data_0083d0.h
@@ -84,7 +84,7 @@ extern u32 var80062914;
 extern s32 g_NumPlayers;
 extern u8 var80062940;
 extern u8 var80062944;
-extern u32 var80062948;
+extern u8 var80062948;
 extern u32 var80062950;
 extern bool g_WeatherTickEnabled;
 extern u32 var80062960;
